requirements.txt
requests
python-telegram-bot
PyQt5
matplotlib
Flask
SQLAlchemy
